[easy] rename symlink test node module
i want to be confident that we're not pulling in one of the other bar_lib modules, so renamed the one we're looking for to something unique.

Differential Revision: D2180715
