+++ tests/node_tests/path_node_modules_with_symlinks/foo.js
@@ -1,2 +1,2 @@
-var x = require('bar_lib');
+var x = require('symlink_lib');
 console.log(x);
