--- tests/node_tests/path_node_modules_with_symlinks/real_node_modules/bar_lib/package.json
@@ -1,1 +1,0 @@
-{"name": "bar_lib", "main": "./index.js"}
