+++ tests/node_tests/path_node_modules_with_symlinks/real_node_modules/symlink_lib/index.js
