+++ tests/node_tests/path_node_modules_with_symlinks/real_node_modules/symlink_lib/package.json
@@ -1,0 +1,1 @@
+{"name": "symlink_lib", "main": "./index.js"}
